+++ packages/sdk/src/utils/Utils.ts
@@ -31,36 +31,34 @@ export class Utils {
 
     let delimiter = "?"; // initial delimiter
 
-    const cicsPlex = (options && options.cicsPlex) == null ? "" : `${encodeURIComponent(options.cicsPlex)}/`;
-    const region = (options && options.regionName) == null ? "" : encodeURIComponent(options.regionName);
+    const cicsPlex = options?.cicsPlex == null ? "" : `${encodeURIComponent(options.cicsPlex)}/`;
+    const region = options?.regionName == null ? "" : encodeURIComponent(options.regionName);
 
     let cmciResource = `/${CicsCmciConstants.CICS_SYSTEM_MANAGEMENT}/${resourceName}/${cicsPlex}${region}`;
 
-    if (options) {
-      if (options.criteria) {
-        cmciResource += `${delimiter}${CicsCmciConstants.CRITERIA}=${this.enforceParentheses(encodeURIComponent(options.criteria))}`;
-        delimiter = "&";
-      }
+    if (options?.criteria) {
+      cmciResource += `${delimiter}${CicsCmciConstants.CRITERIA}=${this.enforceParentheses(encodeURIComponent(options.criteria))}`;
+      delimiter = "&";
+    }
 
-      if (options.parameter) {
-        cmciResource += `${delimiter}PARAMETER=${encodeURIComponent(options.parameter)}`;
-        delimiter = "&";
-      }
+    if (options?.parameter) {
+      cmciResource += `${delimiter}PARAMETER=${encodeURIComponent(options.parameter)}`;
+      delimiter = "&";
+    }
 
-      if (options.queryParams && options.queryParams.summonly) {
-        cmciResource += `${delimiter}${CicsCmciConstants.SUMM_ONLY}`;
-        delimiter = "&";
-      }
+    if (options?.queryParams?.summonly) {
+      cmciResource += `${delimiter}${CicsCmciConstants.SUMM_ONLY}`;
+      delimiter = "&";
+    }
 
-      if (options.queryParams && options.queryParams.nodiscard) {
-        cmciResource += `${delimiter}${CicsCmciConstants.NO_DISCARD}`;
-        delimiter = "&";
-      }
+    if (options?.queryParams?.nodiscard) {
+      cmciResource += `${delimiter}${CicsCmciConstants.NO_DISCARD}`;
+      delimiter = "&";
+    }
 
-      if (options.queryParams && options.queryParams.overrideWarningCount) {
-        cmciResource += `${delimiter}${CicsCmciConstants.OVERRIDE_WARNING_COUNT}`;
-        delimiter = "&";
-      }
+    if (options?.queryParams?.overrideWarningCount) {
+      cmciResource += `${delimiter}${CicsCmciConstants.OVERRIDE_WARNING_COUNT}`;
+      delimiter = "&";
     }
 
     return cmciResource;
@@ -71,7 +69,7 @@ export class Utils {
 
     let cmciResource = `/${CicsCmciConstants.CICS_SYSTEM_MANAGEMENT}/${CicsCmciConstants.CICS_RESULT_CACHE}/${cacheToken}`;
 
-    if (options && options.startIndex) {
+    if (options?.startIndex) {
       cmciResource += `/${options.startIndex}`;
 
       if (options.count) {
@@ -87,7 +85,7 @@ export class Utils {
       delimiter = "&";
     }
 
-    if (options && options.summonly) {
+    if (options?.summonly) {
       cmciResource += `${delimiter}${CicsCmciConstants.SUMM_ONLY}`;
     }